EmissionSystem 추가 및 폭탄 탄착 지점 표시 기능 구현중

+++ Assets/Scripts/Taeseung/Weapon/ParentClass/WeaponManager/SubWeaponManager.cs
@@ -13,7 +13,7 @@ public class SubWeaponManager : WeaponSystem
 
     [Space]
     [Header("WEAPON SUB DETAIL")]
-    [SerializeField] protected int _weaponCount;
+    [SerializeField] protected short _weaponCount;
     [SerializeField] protected float _weaponRemainTime;
